refactor : CustomerJoinRequest 수정
Customer 생성 시 UUID(식별자)는 비즈니스 로직에서 생성합니다.
