+++ src/main/java/com/voucher/vouchermanagement/dto/customer/CustomerJoinRequest.java
@@ -1,0 +1,11 @@
+package com.voucher.vouchermanagement.dto.customer;
+
+public class CustomerJoinRequest {
+    private final String name;
+    private final String email;
+
+    public CustomerJoinRequest(String name, String email) {
+        this.name = name;
+        this.email = email;
+    }
+}
